+++ src/main/java/com/raoulvdberge/refinedstorage/api/network/grid/IGrid.java
@@ -26,6 +26,7 @@ public interface IGrid {
     int SORTING_TYPE_NAME = 1;
     int SORTING_TYPE_ID = 2;
     int SORTING_TYPE_INVENTORYTWEAKS = 3;
+    int SORTING_TYPE_LAST_MODIFIED = 4;
 
     int SEARCH_BOX_MODE_NORMAL = 0;
     int SEARCH_BOX_MODE_NORMAL_AUTOSELECTED = 1;
@@ -232,7 +233,11 @@ public interface IGrid {
     }
 
     static boolean isValidSortingType(int type) {
-        return type == SORTING_TYPE_QUANTITY || type == SORTING_TYPE_NAME || type == SORTING_TYPE_ID || type == SORTING_TYPE_INVENTORYTWEAKS;
+        return type == SORTING_TYPE_QUANTITY ||
+            type == SORTING_TYPE_NAME ||
+            type == SORTING_TYPE_ID ||
+            type == SORTING_TYPE_INVENTORYTWEAKS ||
+            type == SORTING_TYPE_LAST_MODIFIED;
     }
 
     static boolean isValidSortingDirection(int direction) {